# Lib1 performance comparison tests with different codecs