```
Codec: ffv1
Loading 10 Files took 1241 ms (124 ms / frame)
File size reduced from 2.3 MB to 1.9 MB (83.3%)

Codec: ffv1
Loading 100 Files took 6059 ms (60 ms / frame)
File size reduced from 23.6 MB to 18.3 MB (77.6%)

Codec: ffv1
Loading 500 Files took 29272 ms (58 ms / frame)
File size reduced from 113.2 MB to 86.1 MB (76.1%)

Codec: ffv1
Loading 800 Files took 45525 ms (56 ms / frame)
File size reduced from 173.3 MB to 133.4 MB (77.0%)

Codec: ffv1
Loading 1000 Files took 58758 ms (58 ms / frame)
File size reduced from 220.6 MB to 169.7 MB (76.9%)


Codec: vp9
Loading 10 Files took 1804 ms (180 ms / frame)
File size reduced from 2.7 MB to 1.5 MB (54.2%)

Codec: vp9
Loading 100 Files took 17195 ms (171 ms / frame)
File size reduced from 24 MB to 13.7 MB (57.1%)

Codec: vp9
Loading 500 Files took 88192 ms (176 ms / frame)
File size reduced from 114.6 MB to 63 MB (55.0%)

Codec: vp9
Loading 800 Files took 138997 ms (173 ms / frame)
File size reduced from 179.1 MB to 102.6 MB (57.3%)

Codec: vp9
Loading 1000 Files took 164804 ms (164 ms / frame)
File size reduced from 225.4 MB to 129.7 MB (57.6%)
```