@akriese 2023-12-05
Getting similar results, looks fine to me