@akriese 2023-12-05
Please add an info to `VideoGenerator` on how to run the tests in `library/src/main/kotlin` from the command line, I only did it in IntelliJ